result_ctors.rs hands its heap-Ok and Option ctor half to result_ctors_b.rs, a pure text move under the 800-line ceiling
