+++ crates/almide-mir/src/lower/control_b.rs
@@ -431,6 +431,7 @@ include!("heap_result_arm_b.rs");
 include!("heap_result_ctrl_arms.rs");
 include!("result_materialize.rs");
 include!("result_ctors.rs");
+include!("result_ctors_b.rs");
 include!("scalar_for.rs");
 // The defunc HOF family (formerly one 3.5k-line control_p5.rs), split by concern:
 include!("defunc_hof.rs");
