+++ crates/almide-mir/src/lower/result_ctors.rs
@@ -535,472 +535,4 @@ impl LowerCtx {
         self.value_shapes.insert(obj, crate::lower::VariantShape::ResultScalar);
         Some(obj)
     }
-
-    /// Is `ty` a `Result[T_heap, <user variant>]` — the HEAP-Ok structured-error shape
-    /// (`classify: Result[String, MathError]`)? Cap-as-tag (both arms heap), so the
-    /// reader is `seed_variant_param`'s both-heap branch and the eq is
-    /// `result_eq_general_from_handles`'s tag@16 route.
-    pub(crate) fn is_heap_ok_variant_err_result(&self, ty: &Ty) -> bool {
-        use almide_lang::types::constructor::TypeConstructorId;
-        matches!(ty, Ty::Applied(TypeConstructorId::Result, a)
-            if a.len() == 2
-                && is_heap_ty(&a[0])
-                && self.custom_variant_type_name(&a[1]).is_some())
-    }
-
-    /// `err(<user-variant ctor>)` for `Result[T_heap, <user variant>]` — the HEAP-Ok
-    /// twin of [`Self::try_lower_result_err_variant_ctor`]. The variant ctor is INLINED
-    /// (`try_lower_variant_ctor` — a ctor is NOT a wasm fn; the generic heap-Err arm's
-    /// `lower_result_str_piece` Named-call fallback emitted a dangling `(call $NegativeInput)`)
-    /// and MOVED into the CAP-AS-TAG wrapper (`materialize_result_str` — payload @12,
-    /// tag @16 = 1), the exact layout the both-heap reader + `result_eq_general` read.
-    /// A RICH variant type (Overflow(String) — nested heap) routes the wrapper's drop to
-    /// the ERR-side recursion (`reserr:<V>` → `DropWrapperRec` `err_rec`: tag@16 == 1 →
-    /// `$__drop_<V>`, Ok → flat `rc_dec` of the @12 String); a FLAT variant keeps the
-    /// flat `DropListStr` (its block owns no nested heap). `None` outside the shape.
-    pub(crate) fn try_lower_result_err_variant_ctor_heap_ok(
-        &mut self,
-        expr: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        use almide_lang::types::constructor::TypeConstructorId;
-        let err_ty = match result_ty {
-            Ty::Applied(TypeConstructorId::Result, a)
-                if a.len() == 2 && is_heap_ty(&a[0]) =>
-            {
-                &a[1]
-            }
-            _ => return None,
-        };
-        let type_name = self.custom_variant_type_name(err_ty)?;
-        let repr = repr_of(result_ty).ok()?;
-        let IrExprKind::ResultErr { expr: inner } = &expr.kind else {
-            return None;
-        };
-        let piece = self.try_lower_variant_ctor(inner)?;
-        let needs_rec = self.variant_layouts.needs_recursive_drop(&type_name, &|rn| {
-            crate::lower::canonical_record_key(&self.record_layouts, rn).is_some()
-        });
-        // The variant block is MOVED into the wrapper @12 — detach its own scope-end
-        // drop so it frees exactly once, through the wrapper.
-        self.value_drops.get_mut(&piece).map(|d| d.named_route = None);
-        self.value_drops.get_mut(&piece).map(|d| d.flat_elems = false);
-        self.live_heap_handles.retain(|h| *h != piece);
-        let obj = self.materialize_result_str(piece, repr, true, false);
-        if needs_rec {
-            self.value_drops.get_mut(&obj).map(|d| d.flat_elems = false);
-            self.value_drops.entry(obj).or_default().named_route = Some(format!("reserr:{type_name}"));
-        }
-        Some(obj)
-    }
-
-    /// `ok(<Option[R] value>)` / `ok(none)` / `err(<String>)` for `Result[Option[R], String]` where R is
-    /// a record needing a recursive drop — read_message's `ok(none)` / `ok(r)` bases (r:
-    /// `Option[JsonRpcRequest]`). The Ok payload (an Option Var → `Dup`; `some(record)` / `none` →
-    /// materialized) is MOVED into the Result block @12; the wrapper's drop routes to `$__drop_opt_<R>`
-    /// via `resrec:opt_<R>` ([`Op::DropWrapperRec`], certificate UNIFORM over `drop_fn` — no Coq change).
-    /// `$__drop_opt_<R>` is GENERATED (`generate_record_drop_sources`) as `fn __drop_opt_<R>(e: Option[R])
-    /// = match e { some(r) => (), none => () }` (frees the record via `$__drop_<R>` + the Option block).
-    /// `None` outside `Result[Option[<recursive-drop record>], String]` or a non-materializable payload.
-    pub(crate) fn try_lower_result_option_ctor(
-        &mut self,
-        expr: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        use almide_lang::types::constructor::TypeConstructorId;
-        let ok_ty = match result_ty {
-            Ty::Applied(TypeConstructorId::Result, a)
-                if a.len() == 2 && matches!(a[1], Ty::String) =>
-            {
-                &a[0]
-            }
-            _ => return None,
-        };
-        let rec = match ok_ty {
-            Ty::Applied(TypeConstructorId::Option, oa) if oa.len() == 1 => &oa[0],
-            _ => return None,
-        };
-        let rec_drop = self.record_or_anon_drop_type_name(rec)?;
-        let drop_fn = format!("opt_{rec_drop}");
-        let repr = repr_of(result_ty).ok()?;
-        match &expr.kind {
-            IrExprKind::ResultOk { expr: inner } => {
-                let piece = self.lower_option_piece(inner, rec)?;
-                Some(self.materialize_result_aggregate(piece, repr, false, drop_fn))
-            }
-            IrExprKind::ResultErr { expr: inner } => {
-                let piece = self.lower_result_str_piece(inner)?;
-                Some(self.materialize_result_aggregate(piece, repr, true, drop_fn))
-            }
-            _ => None,
-        }
-    }
-
-    /// `ok(some(x))` / `ok(none)` / `err(msg)` RETURNED for a `Result[Option[T], String]` whose Option
-    /// payload is a STRING or a SCALAR leaf (Int/Float/Bool) — the derived-Codec `__decode_option_T`
-    /// shape (`Result[Option[Int], String]` … `Result[Option[String], String]`). The record/tuple/value
-    /// Option payloads are handled by [`Self::try_lower_result_option_ctor`] (recursive `$__drop_opt_<R>`)
-    /// and MUST be left to it — this helper declines them.
-    ///
-    /// The Ok payload is the 0-or-1 Option block (`try_lower_option_ctor` — a scalar `Init::OptSome` or a
-    /// String-holding `DynListStr`), MOVED into the Result @12. The DROP differs by leaf:
-    ///   • SCALAR leaf — the Option[scalar] block owns no inner heap, so the FLAT `materialize_result_str`
-    ///     (`heap_elem_lists` → `DropListStr` `rc_dec`s @12) frees it fully, exactly like a `Result[String,
-    ///     String]`. No generated drop fn.
-    ///   • STRING leaf — the Option[String] block owns the inner String, so a flat `rc_dec` of @12 would
-    ///     LEAK it. Route through `materialize_result_aggregate` with `resrec:opt_str` → the generated
-    ///     `$__drop_opt_str(e: Option[String])` (emitted by `generate_record_drop_sources`), whose
-    ///     `match e { some(r) => (), none => () }` drops the inner String at the some-arm end.
-    pub(crate) fn try_lower_result_option_scalar_str_ctor(
-        &mut self,
-        expr: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        let (ok_ty, is_str) = Self::option_leaf_str_or_scalar(result_ty)?;
-        let repr = repr_of(result_ty).ok()?;
-        match &expr.kind {
-            IrExprKind::ResultOk { expr: inner } => {
-                if is_str {
-                    let opt_repr = repr_of(ok_ty).ok()?;
-                    // Build the `Option[String]` block DIRECTLY: `some(<string>)` co-owns its payload by
-                    // `lower_owned_heap_field` (a Dup for a borrowed param / match-ok String, an Alloc for
-                    // a literal, a move for a call) — `try_lower_option_ctor` declines a borrowed-Var
-                    // payload. `none` is a 0-element block.
-                    let piece = match &inner.kind {
-                        IrExprKind::OptionSome { expr: payload } => {
-                            let s = self.lower_owned_heap_field(payload)?;
-                            self.materialize_opt_str_some(s, opt_repr)
-                        }
-                        IrExprKind::OptionNone => self.materialize_opt_str_none(opt_repr),
-                        _ => return None,
-                    };
-                    // `materialize_opt_str_some`/`_none` mark the block for a flat scope-end `DropListStr`
-                    // (`heap_elem_lists`). It is MOVED into the Result @12 (Consumed) and freed by the
-                    // Result's `resrec:opt_str` → `$__drop_opt_str` instead — detach it so it is freed
-                    // EXACTLY once (no double-free).
-                    self.value_drops.get_mut(&piece).map(|d| d.flat_elems = false);
-                    Some(self.materialize_result_aggregate(piece, repr, false, "opt_str".to_string()))
-                } else {
-                    let piece = self.try_lower_option_ctor(inner, ok_ty)?;
-                    Some(self.materialize_result_str(piece, repr, false, false))
-                }
-            }
-            IrExprKind::ResultErr { expr: inner } => {
-                let piece = self.lower_result_str_piece(inner)?;
-                if is_str {
-                    Some(self.materialize_result_aggregate(piece, repr, true, "opt_str".to_string()))
-                } else {
-                    Some(self.materialize_result_str(piece, repr, true, false))
-                }
-            }
-            _ => None,
-        }
-    }
-
-    /// Build the `Option[R]` Ok payload for [`try_lower_result_option_ctor`]: an Option Var (`Dup` a
-    /// fresh owned ref), `some(record)` (materialize the record into the 0-or-1 Option block via
-    /// `materialize_opt_aggregate_some`), or `none` (a 0-element Option block). `None` otherwise.
-    fn lower_option_piece(&mut self, inner: &IrExpr, rec: &Ty) -> Option<ValueId> {
-        match &inner.kind {
-            // `ok(r)` where r is an owned/borrowed `Option[R]` local — `Dup` a fresh owned reference
-            // (the original drops once at its scope; the Result's @12 owns the Dup'd one).
-            IrExprKind::Var { id } => {
-                let src = self.value_for(*id).ok()?;
-                let dst = self.fresh_value();
-                self.ops.push(Op::Dup { dst, src });
-                Some(dst)
-            }
-            IrExprKind::OptionNone => {
-                // A 0-element Option block (no record inside) — the same empty 0-or-1 layout the
-                // some-builder emits, so `$__drop_opt_<R>` frees it uniformly (its `match` takes none).
-                let repr = repr_of(&inner.ty).ok()?;
-                let z = self.fresh_value();
-                self.ops.push(Op::ConstInt { dst: z, value: 0 });
-                let obj = self.fresh_value();
-                self.ops
-                    .push(Op::Alloc { dst: obj, repr, init: crate::Init::DynListStr { len: z } });
-                self.value_drops.entry(obj).or_default().named_route = Some(format!("opt_{}", self.record_or_anon_drop_type_name(rec)?));
-                Some(obj)
-            }
-            IrExprKind::OptionSome { expr: rec_expr } => {
-                let repr = repr_of(&inner.ty).ok()?;
-                let piece = self.lower_result_str_piece(rec_expr)?;
-                let drop_fn = self.record_or_anon_drop_type_name(rec)?;
-                Some(self.materialize_opt_aggregate_some(piece, repr, drop_fn))
-            }
-            _ => None,
-        }
-    }
-
-    /// Construct a `Result[<non-heap>, String]` `ok(<scalar/unit>)` block — the porta
-    /// `run_foreground` / `ensure_porta_dir` `ok(())` tail and any `ok(<Int/Bool>)`. The Ok payload
-    /// is a SCALAR (or Unit → a `0` placeholder; the @4 len-0 field is the Ok tag consumers read, the
-    /// @12 payload slot is never extracted for a Unit Ok), wrapped by `materialize_result_ok` into the
-    /// flat len-0 block (scope-end `DropListStr` frees just the block — no nested heap to recurse).
-    /// Returns the block (NOT Consumed — the caller moves it out as a tail return, or pushes
-    /// `Op::Consume` for a heap-result-if/match arm). `None` outside `Result[<non-heap>, String]`, a
-    /// non-`ResultOk`, or a HEAP Ok payload — those route to the heap-ok / record / value ctors above.
-    pub(crate) fn try_lower_result_scalar_ok_ctor(
-        &mut self,
-        expr: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        use almide_lang::types::constructor::TypeConstructorId;
-        let ok_ty = match result_ty {
-            Ty::Applied(TypeConstructorId::Result, a)
-                if a.len() == 2 && matches!(a[1], Ty::String) =>
-            {
-                &a[0]
-            }
-            _ => return None,
-        };
-        if is_heap_ty(ok_ty) {
-            return None;
-        }
-        let IrExprKind::ResultOk { expr: inner } = &expr.kind else {
-            return None;
-        };
-        if is_heap_ty(&inner.ty) {
-            return None;
-        }
-        let payload = if matches!(&inner.kind, IrExprKind::Unit) {
-            let z = self.fresh_value();
-            self.ops.push(Op::ConstInt { dst: z, value: 0 });
-            z
-        } else {
-            self.lower_scalar_value(inner)?
-        };
-        let repr = repr_of(result_ty).ok()?;
-        Some(self.materialize_result_ok(payload, repr))
-    }
-
-    /// Construct a `Result[Value, String]` `ok(<Value>)` / `err(<String>)` (the `ok(value.array(...))`
-    /// shape) — the len-1 + tag@16 block, Ok payload a Value (materialized via `lower_owned_heap_field`,
-    /// which handles the `value.*` ctor + nested `list.map`), Err a String. Marked
-    /// `value_result_results` so the drop is the recursive `Op::DropResultValue`. Returns the block
-    /// (NOT yet Consumed — the caller moves it out as a tail return or an arm `Consume`). `None` for a
-    /// non-`Result[Value, String]` type or a payload outside the materializable subset.
-    pub(crate) fn try_lower_result_value_ctor(
-        &mut self,
-        expr: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        if !crate::lower::is_value_result_ty(result_ty) {
-            return None;
-        }
-        let repr = repr_of(result_ty).ok()?;
-        match &expr.kind {
-            IrExprKind::ResultOk { expr: inner } => {
-                let piece = self.lower_owned_heap_field(inner)?;
-                Some(self.materialize_result_str(piece, repr, false, true))
-            }
-            IrExprKind::ResultErr { expr: inner } => {
-                let piece = self.lower_result_str_piece(inner)?;
-                Some(self.materialize_result_str(piece, repr, true, true))
-            }
-            _ => None,
-        }
-    }
-
-    /// Construct a `Result[(String, Int), String]` `ok((<String>, <Int>))` / `err(<String>)` — the
-    /// toml `parse_key_part` `ok((slice, pos))` shape. Ok materializes the `(String, Int)` tuple
-    /// (`try_lower_tuple_construct`, rc-owning the String slot) and wraps it in the cap-as-tag block
-    /// (payload @12 = the tuple handle); Err wraps a String. Tracked in `str_int_result_results` so the
-    /// scope-end drop is the recursive [`Op::DropResultStrInt`] (frees the tuple's String + both blocks)
-    /// — NOT the flat `heap_elem_lists`/`DropListStr` `materialize_result_str` defaults to, which would
-    /// leak the tuple's String. Returns the wrapper block (moved out as the tail return), or `None`
-    /// outside the exact `Result[(String, Int), String]` / materializable-payload subset.
-    pub(crate) fn try_lower_result_str_int_ctor(
-        &mut self,
-        expr: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        use almide_lang::types::constructor::TypeConstructorId;
-        let is_str_int = matches!(result_ty,
-            Ty::Applied(TypeConstructorId::Result, a) if a.len() == 2
-                && matches!(&a[0], Ty::Tuple(ts) if ts.len() == 2
-                    && matches!(ts[0], Ty::String) && matches!(ts[1], Ty::Int))
-                && matches!(a[1], Ty::String));
-        if !is_str_int {
-            return None;
-        }
-        let repr = repr_of(result_ty).ok()?;
-        let obj = match &expr.kind {
-            IrExprKind::ResultOk { expr: inner } => match &inner.kind {
-                IrExprKind::Tuple { elements } => {
-                    let tup = self.try_lower_tuple_construct(elements)?;
-                    self.materialize_result_str(tup, repr, false, false)
-                }
-                _ => return None,
-            },
-            IrExprKind::ResultErr { expr: inner } => {
-                let piece = self.lower_result_str_piece(inner)?;
-                self.materialize_result_str(piece, repr, true, false)
-            }
-            _ => return None,
-        };
-        // Re-route the drop: materialize_result_str(value_ok=false) tracked `heap_elem_lists`
-        // (flat DropListStr); a (String, Int)-tuple Ok needs the recursive DropResultStrInt.
-        self.value_drops.get_mut(&obj).map(|d| d.flat_elems = false);
-        self.value_drops.entry(obj).or_default().str_int_result = true;
-        Some(obj)
-    }
-
-    /// Construct a `Result[(Value, Int), String]` `ok((<Value>, <Int>))` / `err(<String>)` — the toml
-    /// `parse_val` shape. Identical to `try_lower_result_str_int_ctor` except the Ok-tuple's slot0 is a
-    /// dynamic `Value` (so the scope-end drop is the recursive `Op::DropResultValueInt` via
-    /// `$__drop_value_tuple`, tracked in `value_int_result_results`). `None` outside the exact
-    /// `Result[(Value, Int), String]` / materializable-payload subset.
-    pub(crate) fn try_lower_result_value_int_ctor(
-        &mut self,
-        expr: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        if !crate::lower::is_value_int_result_ty(result_ty) {
-            return None;
-        }
-        let repr = repr_of(result_ty).ok()?;
-        let obj = match &expr.kind {
-            IrExprKind::ResultOk { expr: inner } => match &inner.kind {
-                IrExprKind::Tuple { elements } => {
-                    let tup = self.try_lower_tuple_construct(elements)?;
-                    self.materialize_result_str(tup, repr, false, false)
-                }
-                _ => return None,
-            },
-            IrExprKind::ResultErr { expr: inner } => {
-                let piece = self.lower_result_str_piece(inner)?;
-                self.materialize_result_str(piece, repr, true, false)
-            }
-            _ => return None,
-        };
-        self.value_drops.get_mut(&obj).map(|d| d.flat_elems = false);
-        self.value_drops.entry(obj).or_default().value_int_result = true;
-        Some(obj)
-    }
-
-    /// Construct a `Result[(List[Value], Int), String]` `ok((<List[Value]>, <Int>))` / `err(<String>)`
-    /// — toml `collect_array_items`. The Ok-tuple's slot0 is a `List[Value]`, so the scope-end drop is
-    /// the recursive `Op::DropResultListValueInt` (`$__drop_list_value_tuple`), tracked in
-    /// `list_value_int_result_results`. `None` outside the exact type / materializable-payload subset.
-    pub(crate) fn try_lower_result_list_value_int_ctor(
-        &mut self,
-        expr: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        if !crate::lower::is_list_value_int_result_ty(result_ty) {
-            return None;
-        }
-        let repr = repr_of(result_ty).ok()?;
-        let obj = match &expr.kind {
-            IrExprKind::ResultOk { expr: inner } => match &inner.kind {
-                IrExprKind::Tuple { elements } => {
-                    let tup = self.try_lower_tuple_construct(elements)?;
-                    self.materialize_result_str(tup, repr, false, false)
-                }
-                _ => return None,
-            },
-            IrExprKind::ResultErr { expr: inner } => {
-                let piece = self.lower_result_str_piece(inner)?;
-                self.materialize_result_str(piece, repr, true, false)
-            }
-            _ => return None,
-        };
-        self.value_drops.get_mut(&obj).map(|d| d.flat_elems = false);
-        self.value_drops.entry(obj).or_default().list_value_int_result = true;
-        Some(obj)
-    }
-
-    /// Construct a `Result[(List[String], Int), String]` `ok((<List[String]>, <Int>))` / `err(<String>)`
-    /// — the toml `parse_key` / `parse_table_key` shape. The Ok-tuple's slot0 is a `List[String]`, so
-    /// the scope-end drop is the recursive `Op::DropResultListStrInt` (frees each element String + the
-    /// List block + the tuple block), tracked in `list_str_int_result_results`. `None` outside the exact
-    /// `Result[(List[String], Int), String]` / materializable-payload subset.
-    pub(crate) fn try_lower_result_list_str_int_ctor(
-        &mut self,
-        expr: &IrExpr,
-        result_ty: &Ty,
-    ) -> Option<ValueId> {
-        if !crate::lower::is_list_str_int_result_ty(result_ty) {
-            return None;
-        }
-        let repr = repr_of(result_ty).ok()?;
-        let obj = match &expr.kind {
-            IrExprKind::ResultOk { expr: inner } => match &inner.kind {
-                IrExprKind::Tuple { elements } => {
-                    let tup = self.try_lower_tuple_construct(elements)?;
-                    self.materialize_result_str(tup, repr, false, false)
-                }
-                _ => return None,
-            },
-            IrExprKind::ResultErr { expr: inner } => {
-                let piece = self.lower_result_str_piece(inner)?;
-                self.materialize_result_str(piece, repr, true, false)
-            }
-            _ => return None,
-        };
-        self.value_drops.get_mut(&obj).map(|d| d.flat_elems = false);
-        self.value_drops.entry(obj).or_default().list_str_int_result = true;
-        Some(obj)
-    }
-
-    /// `handle + k` as a fresh i64 address value (ConstInt + IntBinOp::Add).
-    fn const_add(&mut self, base: ValueId, k: i64) -> ValueId {
-        let c = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: c, value: k });
-        let dst = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst, op: IntOp::Add, a: base, b: c });
-        dst
-    }
-
-    /// `Ok(int)` for `Result[Int, String]` = a cap-1/len-0 `DynListStr`: allocate ONE element slot
-    /// (so the block is the same physical size as an `Err`'s, free-list-compatible via cap), store
-    /// the int in slot 0, then OVERRIDE the len field to 0 so `DropListStr` frees no element (the
-    /// int is scalar, owns nothing). Cert: a `None`-like DynListStr (Alloc `i`, no String move-in,
-    /// scope-end DropListStr `d`) — the int store + len override are opaque prim ops the checker
-    /// ignores. The tag read (len == 0) marks it `Ok`.
-    /// `Err(<scalar>)` for a SCALAR-SCALAR Result (`Result[Int, Int]` — the
-    /// match_container `ck(err(404))` class): the SAME len-as-tag block as
-    /// [`Self::materialize_result_ok`] but the len field STAYS 1 (the Err tag) and slot 0
-    /// holds the SCALAR err payload. Deliberately NOT `heap_elem_lists`-tracked: a
-    /// DropListStr over len 1 would rc_dec the raw scalar as a handle (the rc_dec-trap
-    /// class); the caller's flat `Op::Drop` frees the block exactly (neither arm owns
-    /// children). Cert: one Alloc `i` + the scope-end `d` — the same balanced pair.
-    pub(crate) fn materialize_result_err_scalar(
-        &mut self,
-        payload: ValueId,
-        repr: crate::Repr,
-    ) -> ValueId {
-        use crate::PrimKind;
-        let one = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: one, value: 1 });
-        let obj = self.fresh_value();
-        self.ops.push(Op::Alloc { dst: obj, repr, init: Init::DynList { len: one } });
-        let oh = self.fresh_value();
-        self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(oh), args: vec![obj] });
-        // slot 0 (handle + 12) = the scalar Err payload; len stays 1 = the Err tag.
-        let twelve = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: twelve, value: 12 });
-        let daddr = self.fresh_value();
-        self.ops.push(Op::IntBinOp { dst: daddr, op: IntOp::Add, a: oh, b: twelve });
-        self.ops.push(Op::Prim {
-            kind: PrimKind::Store { width: 8 },
-            dst: None,
-            args: vec![daddr, payload],
-        });
-        obj
-    }
-
-    /// `ok(<scalar>)` — ONE semantic `Alloc { init: ResOkScalar }` (the
-    /// result-family-from-type "desugar once" slice). The wasm render expands it
-    /// to the byte-identical len-as-tag block the old 6-op window built (len 0 =
-    /// Ok tag, payload @12, @16 zeroed); the native leg maps it 1:1 onto
-    /// `PrimKind::ResMakeOk` in native_result_rewrite — a total single-op match
-    /// that replaced the fragile producer-window recognition.
-    pub(crate) fn materialize_result_ok(&mut self, payload: ValueId, repr: crate::Repr) -> ValueId {
-        let obj = self.fresh_value();
-        self.ops.push(Op::Alloc { dst: obj, repr, init: Init::ResOkScalar { payload } });
-        self.value_drops.entry(obj).or_default().flat_elems = true;
-        obj
-    }
 }
